% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_webr.R
\name{reprex_webr}
\alias{reprex_webr}
\title{Create a webR Share URL from R code}
\usage{
reprex_webr(
  x = NULL,
  input = NULL,
  base_url = "https://webr.r-wasm.org/latest/",
  html_preview = TRUE
)
}
\arguments{
\item{x}{An expression. If not provided, will try \code{input}, then clipboard.}

\item{input}{Character. If length 1 and a file exists at that path, code is read from file. 
Otherwise, assumed to be code as character vector.}

\item{base_url}{The base webR URL for sharing (default: "https://webr.r-wasm.org/latest/").}

\item{html_preview}{Logical. Whether to show rendered output in a viewer (RStudio or browser). Always FALSE in a noninteractive session.}
}
\value{
A single string containing the shareable URL.
}
\description{
Prepares R code for sharing and execution in a webR browser session by 
encoding it into a URL. Accepts code as an expression, character vector, 
file path, or from the clipboard, similar to reprex().
}
\examples{
clipr::write_clip('print("hello reprex!")')
reprex_webr()
reprex_webr(input = c("x <- 1:10", "plot(x, x^2)"))
reprex_webr({x <- 1:10; plot(x, x^2)})

}
